refactor(types): split key-rotation.ts into 2 per-entity files
