+++ packages/types/src/api-constants/api-limits.ts
@@ -18,7 +18,8 @@ import {
 } from "./time-constants.js";
 
 import type { BlobPurpose } from "../entities/blob.js";
-import type { RotationItemStatus, RotationState } from "../key-rotation.js";
+import type { RotationState } from "../entities/bucket-key-rotation.js";
+import type { RotationItemStatus } from "../entities/bucket-rotation-item.js";
 
 // ── Pagination ───────────────────────────────────────────────────────
